refactor: GameUi클래스를 InputView와 OutputView로 분리 및 입력값 테스트 파일이름 수정
